requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.4.1
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.4.1
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
blis==1.0.1
bs4==0.0.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.10
dataclasses-json==0.6.7
dateparser==1.2.0
debugpy==1.8.1
decorator==5.1.1
discord==2.3.2
discord.py==2.3.2
distlib==0.3.8
distro==1.9.0
Django==5.0.2
elastic-transport==8.15.1
elasticsearch==8.16.0
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
executing==2.0.1
fake-useragent==1.5.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.13.1
grpcio==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.7
langchain==0.3.13
langchain-core==0.3.28
langchain-groq==0.2.2
langchain-mixedbread-ai==0.1.0
langchain-text-splitters==0.3.4
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.2.7
language_data==1.3.0
lxml==5.3.0
lxml_html_clean==0.4.1
lz4==4.3.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mixedbread-ai==2.2.6
mlxtend==0.23.1
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.11
mypy-extensions==1.0.0
narwhals==1.9.0
nest-asyncio==1.6.0
networkx==3.4.2
newsapi==0.1.1
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.0
oauthlib==3.2.2
openai==1.58.1
orjson==3.10.13
packaging==24.2
pandas==2.2.3
parse==1.20.2
parso==0.8.4
patsy==1.0.1
pillow==11.0.0
pinecone==5.4.2
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.1.0
preshed==3.0.9
prompt-toolkit==3.0.43
propcache==0.2.1
protobuf==4.25.5
protoc-gen-openapiv2==0.0.1
psutil==5.9.8
pure-eval==0.2.2
pyarrow==17.0.0
pycparser==2.21
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pyee==11.1.1
Pygments==2.18.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.0
pypng==0.20220715.0
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.0.3
qrcode==7.4.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-html==0.10.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
safetensors==0.5.0
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.6.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.39.0
sympy==1.13.1
tenacity==9.0.0
thinc==8.3.2
threadpoolctl==3.5.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
tornado==6.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
tweepy==4.14.0
typer==0.14.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
virtualenv==20.25.0
w3lib==2.2.1
wasabi==1.1.3
watchdog==5.0.3
wcwidth==0.2.13
weasel==0.4.1
websockets==10.4
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
youtube-dl==2021.12.17
zipp==3.21.0
